Don't cast pointers to longs in asm statements.
These asm statments are not quite as pessimal as when I complained
about them in rev.1.9 of audio.c.  They seem to be only 40% slower
than the C version on P5's and the same speed on K6's.


Former-commit-id: 46d8f86072db7ff44ded4728c48400639388b7fa

+++ sys/i386/isa/snd/sound.c
@@ -6130,7 +6130,7 @@ operator|>
 literal|0
 condition|)
 block|{
-asm|__asm__ (  "   cld\n" 		   "1: lodsb\n" 		   "   xlatb\n" 		   "   stosb\n" 		   "   loop 1b\n": 	    : "b" ((long) table), "c" (n), "D" ((long) buff), "S" ((long) buff) 	    : "bx", "cx", "di", "si", "ax");
+asm|__asm__ (  "   cld\n" 		   "1: lodsb\n" 		   "   xlatb\n" 		   "   stosb\n" 		   "   loop 1b\n": 	    : "b" (table), "c" (n), "D" (buff), "S" (buff) 	    : "bx", "cx", "di", "si", "ax");
 block|}
 block|}
 end_function
